+++ app/src/main/java/io/github/bropines/tailscaled/appfunctions/TailSocksAppFunctionService.kt
@@ -1,0 +1,5 @@
+package io.github.bropines.tailscaled.appfunctions
+
+import androidx.appfunctions.AppFunctionService
+
+abstract class TailSocksAppFunctionService : AppFunctionService()